Remove RSpec::Core::Formatters for RSpec 2
