--- lib/rspec/core/formatters/turnip_formatter/for_rspec2.rb
@@ -1,42 +1,0 @@
-# -*- coding: utf-8 -*-
-
-module RSpec
-  module Core
-    module Formatters
-      class TurnipFormatter < BaseFormatter
-        module ForRSpec2
-          def dump_summary(duration, _, failure_count, pending_count)
-            print_params = {
-              scenarios:      scenarios,
-              failed_count:   failure_count,
-              pending_count:  pending_count,
-              total_time:     duration
-            }
-            output_html(print_params)
-          end
-
-          def example_passed(example)
-            scenario = ::TurnipFormatter::Scenario::Pass.new(example)
-            scenarios << scenario
-          end
-
-          def example_pending(example)
-            scenario = ::TurnipFormatter::Scenario::Pending.new(example)
-            scenarios << scenario
-          end
-
-          def example_failed(example)
-            scenario = ::TurnipFormatter::Scenario::Failure.new(example)
-            scenarios << scenario
-          end
-
-          module Helper
-            def formatted_backtrace(example)
-              RSpec::Core::BacktraceFormatter.format_backtrace(example.exception.backtrace, example.metadata)
-            end
-          end
-        end
-      end
-    end
-  end
-end
