+++ turnip_formatter.gemspec
@@ -20,7 +20,7 @@ Gem::Specification.new do |spec|
 
   spec.add_dependency 'turnip', '~> 1.3.0'
   spec.add_dependency 'slim'
-  spec.add_dependency 'rspec', ['>=2.14.0', '<4.0']
+  spec.add_dependency 'rspec', [">=3.0", "<3.4"]
   spec.add_development_dependency 'bundler', '~> 1.3'
   spec.add_development_dependency 'rake'
   spec.add_development_dependency 'coveralls'
